Refactor matchers, improve previous
This refactors the matchers to be more reusable.

Additionally, there is a bug in previous. Because previous is not going
to the previous block, but just searching for keywords, it ignores
indentation. This should either search for the next block up moving
around/using searchpair, or at the very least respect the indentation
and use that to figure out next sibling.

Additionally, for next blocks, we should match the block keyword, not
just do.