Update previous block to stop at parent
Previously, if you ran `RubyBlockPrevious` at the first child in a
parent, i.e.:

```
context "foo" do
  ...
end

describe "bar" do
  it "baz" do
```

If your cursor was on the "baz" line, you would go to the "foo" line.
This felt a little unintuitive. Now, previous would take you to the
"bar" line, which in this case is the parent.

This does cause a discrepancy with `RubyBlockNext`. If you were to call
the next function on the last child in a block, this code will continue
to go to the first sibling of the parent. I.e.

```
describe "bar" do
  it "baz" do
  end
end

context "foo" do
  ...
end
```

If you were on the "baz" line, you would end up on the "foo" line. I
still think this should be the behavior only due to the value of the
first/start line of a block vs the `end` that closes the block. I don't
see any value in ending up on "end" vs with the previous block, ending
up on the parent does seem valuable.