Update do regex to ignore extra do's
Previously, if there was a test description (or anything) that had a do
in it, i.e.

```
it "will do the thing" do
end
```

That would cause a variety of issues. This commit updates the do regex
do look for do's that are at the end of the line or only followed by
spaces or have something in pipes, i.e. `do |foo|`.