Adjust visual mode, fix next block when on end
Previously, the visual mappings would always enter visual mode and never
visual line mode. For now, I am defaulting to always be in visual line
mode. If this is a problem I can properly set apart the two mappings.

Additionally, if you are on an `end` line, then the next ruby block
would often not find a sibling, but instead would jump to the parent's
sibling because it goes to the end of a block first. This makes sure it
only goes to block end if it is NOT on an end.