Simplify visual check, fix function scope
The main point of this commit is to fix the scope of the functions in
this plugin. Previously they were globally scoped, and they really
should be specific to the script. I realize that _technically_ this is a
breaking change if someone was using the functions directly, but in the
spirit of semver, this is a breaking change to an API I didn't really
intend to be public. Also, I didn't understand best practices.

Speaking of best practices, this also no longer defines the commands or
functions with a `!` so they won't overwrite existing ones, although
those hopefully don't exist.

Finally, there is small simplification to the visual check.